refactor: integrate asset category filters in dashboard components without duplication
