+++ src/hooks/useCoinGeckoCategories.ts
